+++ cognite/neat/graph/loaders/_base.py
@@ -9,6 +9,7 @@ from cognite.client.data_classes.capabilities import Capability
 from cognite.neat.graph import NeatGraphStore
 from cognite.neat.graph.issues.loader import FailedAuthorizationError
 from cognite.neat.issues import NeatIssue, NeatIssueList
+from cognite.neat.utils.auxiliary import class_html_doc
 from cognite.neat.utils.upload import UploadResult, UploadResultList
 
 T_Output = TypeVar("T_Output")
@@ -33,6 +34,10 @@ class BaseLoader(ABC, Generic[T_Output]):
     def _load(self, stop_on_exception: bool = False) -> Iterable[T_Output | NeatIssue]:
         """Load the graph with data."""
         pass
+
+    @classmethod
+    def _repr_html_(cls) -> str:
+        return class_html_doc(cls)
 
 
 class CDFLoader(BaseLoader[T_Output]):