+++ tests/tests_unit/test_utils/test_auxiliary.py
@@ -1,0 +1,12 @@
+import pytest
+
+from cognite.neat.rules.importers import DMSImporter
+from cognite.neat.utils.auxiliary import get_classmethods
+
+
+@pytest.mark.parametrize(
+    "cls_, expected_methods",
+    [(DMSImporter, [DMSImporter.from_data_model_id, DMSImporter.from_directory, DMSImporter.from_zip_file])],
+)
+def test_get_classmethods(cls_, expected_methods: list) -> None:
+    assert get_classmethods(cls_) == expected_methods